+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- **CheckPartial** icon for new partial state in Checkbox.
+
 ## [8.36.2] - 2019-04-18
 ### Fixed
 - Fix `withForwardedRef` type, remove oneOfType usage.